tests: consistently use `get_krun_add_net_unixgram`, don't duplicate the definition
